# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "YP API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/reportedUsers": {
            "get": {
                "tags": [
                    "ReportedUsers"
                ],
                "summary": "Get list of ReportedUsers",
                "description": "Returns list of ReportedUsers",
                "operationId": "indexReportedUsers",
                "parameters": [
                    {
                        "name": "idUser",
                        "in": "query",
                        "description": "id of the User who ask the request",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "ReportedUsers"
                ],
                "summary": "Report a Users",
                "description": "Returns a ReportedUsers",
                "operationId": "storeReportedUsers",
                "parameters": [
                    {
                        "name": "reason",
                        "in": "query",
                        "description": "reason of the report",
                        "required": true
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "description": "content of the report",
                        "required": true
                    },
                    {
                        "name": "reportedUser",
                        "in": "query",
                        "description": "id of the User who's been reported",
                        "required": true
                    },
                    {
                        "name": "userWhoReported",
                        "in": "query",
                        "description": "id of the User who reported",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/reportedUsers/{id}": {
            "get": {
                "tags": [
                    "ReportedUsers"
                ],
                "summary": "Get the reports from one Users",
                "description": "Returns a ReportedUsers",
                "operationId": "showReportedUsers",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Not used",
                        "required": true
                    },
                    {
                        "name": "reportedUser",
                        "in": "query",
                        "description": "id of the User who's been reported",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "ReportedUsers"
                ],
                "summary": "Delete a Report",
                "description": "Returns a ReportedUsers",
                "operationId": "deleteReportedUsers",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of the report",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "ReportedUsers"
                ],
                "summary": "Update a Report",
                "description": "Returns a ReportedUsers",
                "operationId": "updateReportedUsers",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of the report",
                        "required": true
                    },
                    {
                        "name": "reason",
                        "in": "query",
                        "description": "reason of the report",
                        "required": true
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "description": "content of the report",
                        "required": true
                    },
                    {
                        "name": "reportedUser",
                        "in": "query",
                        "description": "id of the User who's been reported",
                        "required": true
                    },
                    {
                        "name": "userWhoReported",
                        "in": "query",
                        "description": "id of the User who reported",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get user",
                "description": "Returns user",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the user",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete One User",
                "description": "Returns true or false",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the User you want to delete",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "User"
                ],
                "summary": "Update one user",
                "description": "Returns data's of the user",
                "operationId": "update",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}